+++ HISTORY.md
@@ -13,6 +13,9 @@
 
   * Bump minimum required version of Armadillo to 4.200.0.
 
+  * Added simple Gradient Descent optimizer, found in
+    src/mlpack/core/optimizers/gradient_descent/ (#792).
+
 ### mlpack 2.0.3
 ###### 2016-07-21
   * Added multiprobe LSH (#691).  The parameter 'T' to LSHSearch::Search() can